vibe/pyproject.toml
[project]
name = "vibe"
version = "0.1.0"
description = "Task Management Assistant using Strands SDK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents[openai]>=0.1.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]
